+++ contracts/token/erc721/preset/ImmutableERC721.sol
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache 2.0
 pragma solidity 0.8.19;
 
-import {MintingAccessControl} from "../abstract/MintingAccessControl.sol";
 import {ImmutableERC721HybridBase} from "../abstract/ImmutableERC721HybridBase.sol";
 
 contract ImmutableERC721 is ImmutableERC721HybridBase {
@@ -121,7 +120,7 @@ contract ImmutableERC721 is ImmutableERC721HybridBase {
             revert IImmutableERC721MismatchedTransferLengths();
         }
 
-        for (uint i = 0; i < tr.tokenIds.length; i++) {
+        for (uint256 i = 0; i < tr.tokenIds.length; i++) {
             safeTransferFrom(tr.from, tr.tos[i], tr.tokenIds[i]);
         }
     }
